[CHANGE] Moved Discord notification handling to its own module
